backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.0
sqlalchemy>=2.0
psycopg[binary]>=3.1

# RAG & LLM
langchain>=0.1.0
langchain-google-genai>=0.0.7
google-generativeai>=0.3.0
pgvector>=0.3.6

# Async support
httpx>=0.20.0

# CORS
python-multipart>=0.0.5

# Testing (optional)
pytest>=7.0
pytest-asyncio>=0.20
